dom/events/test/bug1017086_inner.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1017086
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1017086</title>
  <meta name="author" content="Maksim Lebedev" />
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">
    /** Test for Bug 1017086 **/
    var testelem = undefined;
    var pointer_events = ["onpointerover", "onpointerenter",
                          "onpointermove",
                          "onpointerdown", "onpointerup",
                          "onpointerout",  "onpointerleave",
                          "onpointercancel"];
    function check(expected_value, event_name, container, container_name) {
      var text = event_name + " in " + container_name + " should be " + expected_value;
      parent.is(event_name in container, expected_value, text);
    }
    function runTest() {
      testelem = document.getElementById("test");
      is(!!testelem, true, "Document should have element with id 'test'");
      parent.turnOnOffPointerEvents( function() {
        parent.part_of_checks(pointer_events, check, window, document, testelem);
      });
    }
  </script>
</head>
<body onload="runTest();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1017086">Mozilla Bug 1017086</a>
  <p id="display"></p>
  <div id="content" style="display: none">
  </div>
  <pre id="test">
  </pre>
</body>
</html>
